fix name collision issue, cleanup
