+++ fogros2/fogros2/launch_description.py
@@ -44,89 +44,12 @@ if TYPE_CHECKING:
         IncludeLaunchDescription,
     )  # noqa: F401
 
-import os
 import pickle
 from collections import defaultdict
 from threading import Thread
 from time import sleep
 
-import wgconfig
-import wgconfig.wgexec as wgexec
-
-
-class VPN:
-    def __init__(
-        self,
-        cloud_key_path="/tmp/fogros-cloud.conf",
-        robot_key_path="/tmp/fogros-local.conf",
-    ):
-        self.cloud_key_path = cloud_key_path
-        self.robot_key_path = robot_key_path
-
-        self.cloud_name_to_pub_key_path = dict()
-        self.cloud_name_to_priv_key_path = dict()
-
-        self.robot_private_key = wgexec.generate_privatekey()
-        self.robot_public_key = wgexec.get_publickey(self.robot_private_key)
-
-    def generate_key_pairs(self, machines):
-        """
-        Create key pair for each machine.
-
-        @param machines: List<machine>
-        """
-        for machine in machines:
-            name = machine.get_name()
-            cloud_private_key = wgexec.generate_privatekey()
-            self.cloud_name_to_priv_key_path[name] = cloud_private_key
-            cloud_public_key = wgexec.get_publickey(cloud_private_key)
-            self.cloud_name_to_pub_key_path[name] = cloud_public_key
-
-    def generate_wg_config_files(self, machines):
-        self.generate_key_pairs(machines)
-
-        # generate cloud configs
-        counter = 2  # start the static ip addr counter from 2
-        for machine in machines:
-            name = machine.get_name()
-            machine_config_pwd = self.cloud_key_path + name
-            machine_priv_key = self.cloud_name_to_priv_key_path[name]
-            aws_config = wgconfig.WGConfig(machine_config_pwd)
-            aws_config.add_attr(None, "PrivateKey", machine_priv_key)
-            aws_config.add_attr(None, "ListenPort", 51820)
-            aws_config.add_attr(
-                None, "Address", "10.0.0." + str(counter) + "/24"
-            )
-            aws_config.add_peer(self.robot_public_key, "# fogROS Robot")
-            aws_config.add_attr(
-                self.robot_public_key, "AllowedIPs", "10.0.0.1/32"
-            )
-            aws_config.write_file()
-            counter += 1
-
-        # generate robot configs
-        robot_config = wgconfig.WGConfig(self.robot_key_path)
-        robot_config.add_attr(None, "PrivateKey", self.robot_private_key)
-        robot_config.add_attr(None, "ListenPort", 51820)
-        robot_config.add_attr(None, "Address", "10.0.0.1/24")
-        for machine in machines:
-            name = machine.get_name()
-            ip = machine.get_ip()
-            cloud_pub_key = self.cloud_name_to_pub_key_path[name]
-            robot_config.add_peer(cloud_pub_key, "# AWS" + name)
-            robot_config.add_attr(cloud_pub_key, "AllowedIPs", "10.0.0.2/32")
-            robot_config.add_attr(cloud_pub_key, "Endpoint", f"{ip}:51820")
-            robot_config.add_attr(cloud_pub_key, "PersistentKeepalive", 3)
-        robot_config.write_file()
-
-    def start_robot_vpn(self):
-        # Copy /tmp/fogros-local.conf to /etc/wireguard/wg0.conf locally.
-        # TODO: This needs root. Move this to a separate script with setuid.
-
-        os.system("sudo cp /tmp/fogros-local.conf /etc/wireguard/wg0.conf")
-        os.system("sudo chmod 600 /etc/wireguard/wg0.conf")
-        os.system("sudo wg-quick down wg0")
-        os.system("sudo wg-quick up wg0")
+from .vpn import VPN
 
 
 class FogROSLaunchDescription(LaunchDescriptionEntity):
@@ -154,7 +77,6 @@ class FogROSLaunchDescription(LaunchDescriptionEntity):
         deprecated_reason: Optional[Text] = None,
     ) -> None:
         """Create a LaunchDescription."""
-        launch.logging.get_logger().info("init")
         self.__entities = []
         self.__to_cloud_entities = defaultdict(list)
         self.__streamed_topics = []
@@ -170,8 +92,8 @@ class FogROSLaunchDescription(LaunchDescriptionEntity):
         """Override LaunchDescriptionEntity to visit contained entities."""
         # dump the to cloud nodes into different files
         for key, value in self.__to_cloud_entities.items():
-            with open("/tmp/to_cloud_" + key, "wb+") as f:
-                print(key + ": to be dumped")
+            with open(f"/tmp/to_cloud_{key}", "wb+") as f:
+                print(f"{key}: to be dumped")
                 dumped_node_str = pickle.dumps(value)
                 f.write(dumped_node_str)
 
@@ -187,9 +109,8 @@ class FogROSLaunchDescription(LaunchDescriptionEntity):
         # tell remote machine to push the to cloud nodes and
         # wait here until all the nodes are done
         for machine in machines:
-            machine_name = machine.get_name()
-            while not machine.get_ready_state():
-                print("Waiting for machine " + machine_name)
+            while not machine.is_created:
+                print(f"Waiting for machine {machine.name}")
                 sleep(1)
             # machine is ready, # push to_cloud and setup vpn
             machine.push_to_cloud_nodes()
@@ -337,7 +258,7 @@ class FogROSLaunchDescription(LaunchDescriptionEntity):
 
     def add_entity_with_filter(self, entity):
         if entity.__class__.__name__ == "CloudNode":
-            self.__to_cloud_entities[entity.get_unique_id()].append(entity)
+            self.__to_cloud_entities[entity.unique_id].append(entity)
             if entity.stream_topics:
                 for stream_topic in entity.stream_topics:
                     self.add_image_transport_entities(
@@ -355,7 +276,7 @@ class FogROSLaunchDescription(LaunchDescriptionEntity):
         import fogros2
 
         self.__streamed_topics.append(topic_name)
-        new_cloud_topic_name = topic_name + "/cloud"
+        new_cloud_topic_name = f"{topic_name}/cloud"
         print(
             f"Added {intermediate_transport} transport decoder/subscriber "
             f"for topic {topic_name}"
@@ -372,16 +293,14 @@ class FogROSLaunchDescription(LaunchDescriptionEntity):
             ],
             remappings=[
                 (
-                    "in/" + intermediate_transport,
-                    topic_name + "/" + intermediate_transport,
+                    f"in/{intermediate_transport}",
+                    f"{topic_name}/{intermediate_transport}",
                 ),
                 ("out", new_cloud_topic_name),
             ],
         )
 
-        self.__to_cloud_entities[decoder_node.get_unique_id()].append(
-            decoder_node
-        )
+        self.__to_cloud_entities[decoder_node.unique_id].append(decoder_node)
 
         print(
             f"Added {intermediate_transport} transport encoder/publisher "
@@ -399,8 +318,8 @@ class FogROSLaunchDescription(LaunchDescriptionEntity):
             remappings=[
                 ("in", topic_name),
                 (
-                    "out/" + intermediate_transport,
-                    topic_name + "/" + intermediate_transport,
+                    f"out/{intermediate_transport}",
+                    f"{topic_name}/{intermediate_transport}",
                 ),
             ],
         )
